more typecast in DBGPRINT

+++ common/spectrum.c
@@ -1846,7 +1846,8 @@ static void PeerMeasureReportAction(
 /*	if ((pMeasureReportInfo = kmalloc(sizeof(MEASURE_RPI_REPORT), GFP_ATOMIC)) == NULL)*/
 	if (pMeasureReportInfo == NULL)
 	{
-		DBGPRINT(RT_DEBUG_ERROR, ("%s unable to alloc memory for measure report buffer (size=%d).\n", __FUNCTION__, sizeof(MEASURE_RPI_REPORT)));
+		DBGPRINT(RT_DEBUG_ERROR, ("%s unable to alloc memory for measure report buffer (size=%d).\n",
+			__FUNCTION__, (int) sizeof(MEASURE_RPI_REPORT)));
 		return;
 	}
